Merge

+++ src/com/android/providers/downloads/DownloadService.java
@@ -260,7 +260,7 @@ public class DownloadService extends Service {
 
         public void run() {
             Process.setThreadPriority(Process.THREAD_PRIORITY_BACKGROUND);
-
+			createCacheDirs();
             trimDatabase();
             removeSpuriousFiles();
 
@@ -363,7 +363,7 @@ public class DownloadService extends Service {
                             if (info.shouldScanFile()) {
                                 // initiate rescan of the file to - which will populate
                                 // mediaProviderUri column in this row
-                                if (!scanFile(info, false, true)) {
+                                if (!scanFile(info, true, false)) {
                                     throw new IllegalStateException("scanFile failed!");
                                 }
                             } else {
@@ -377,9 +377,9 @@ public class DownloadService extends Service {
                             // in DownProvider database (the order of deletion is important).
                             getContentResolver().delete(Uri.parse(info.mMediaProviderUri), null,
                                     null);
-                            // the following deletes the file and then deletes it from downloads db
-                            Helpers.deleteFile(getContentResolver(), info.mId, info.mFileName,
-                                    info.mMimeType);
+                            getContentResolver().delete(Downloads.Impl.ALL_DOWNLOADS_CONTENT_URI,
+                                    Downloads.Impl._ID + " = ? ",
+                                    new String[]{String.valueOf(info.mId)});
                         }
                     }
                 }
@@ -417,7 +417,16 @@ public class DownloadService extends Service {
                             PendingIntent.FLAG_ONE_SHOT));
         }
     }
-
+	
+	/*
+	* Checks for and creates download cache directory
+	* we need this just in case it doesn't exist on the sdcard
+	*/
+	private void createCacheDirs() {
+		File cacheDir = Environment.getDownloadCacheDirectory();
+		cacheDir.mkdirs();
+	} 
+	
     /**
      * Removes files that may have been left behind in the cache directory
      */
@@ -580,23 +589,21 @@ public class DownloadService extends Service {
                 mMediaScannerService.requestScanFile(info.mFileName, info.mMimeType,
                         new IMediaScannerListener.Stub() {
                             public void scanCompleted(String path, Uri uri) {
-                                if (updateDatabase) {
-                                    // Mark this as 'scanned' in the database
-                                    // so that it is NOT subject to re-scanning by MediaScanner
-                                    // next time this database row is encountered
+                                if (uri != null && updateDatabase) {
+                                    // file is scanned and mediaprovider returned uri. store it in downloads
+                                    // table (i.e., update this downloaded file's row)
                                     ContentValues values = new ContentValues();
                                     values.put(Constants.MEDIA_SCANNED, 1);
-                                    if (uri != null) {
-                                        values.put(Downloads.Impl.COLUMN_MEDIAPROVIDER_URI,
-                                                uri.toString());
-                                    }
+                                    values.put(Downloads.Impl.COLUMN_MEDIAPROVIDER_URI,
+                                            uri.toString());
                                     getContentResolver().update(key, values, null, null);
-                                } else if (deleteFile) {
-                                    if (uri != null) {
-                                        // use the Uri returned to delete it from the MediaProvider
-                                        getContentResolver().delete(uri, null, null);
-                                    }
-                                    // delete the file and delete its row from the downloads db
+                                } else if (uri == null && deleteFile) {
+                                    // callback returned NO uri..that means this file doesn't
+                                    // exist in MediaProvider. but it still needs to be deleted
+                                    // TODO don't scan files that are not scannable by MediaScanner.
+                                    //      create a public method in MediaFile.java to return false
+                                    //      if the given file's mimetype is not any of the types
+                                    //      the mediaprovider is interested in.
                                     Helpers.deleteFile(resolver, id, path, mimeType);
                                 }
                             }
